@@ -1,39 +1,45 @@
 import 'package:hive/hive.dart';
 
-part 'unit.g.dart';
+part 'keyword.g.dart';
 
-@HiveType(typeId: 7)
-class Unit {
+@HiveType(typeId: 13)
+class Keyword {
 
   @HiveField(0)
   final int? id;
   @HiveField(1)
-  final String name;
+  final String? name;
   @HiveField(2)
+  final String? label;
+  @HiveField(3)
   final String? description;
 
-  Unit({
+  Keyword({
     this.id,
-    required this.name,
+    this.name,
+    this.label,
     this.description
   });
 
-  Unit copyWith({
+  Keyword copyWith({
     int? id,
     String? name,
+    String? label,
     String? description,
   }) {
-    return Unit(
+    return Keyword(
       id: id ?? this.id,
       name: name ?? this.name,
+      label: label ?? this.label,
       description: description ?? this.description,
     );
   }
 
-  factory Unit.fromJson(Map<String, dynamic> json) {
-    return Unit(
+  factory Keyword.fromJson(Map<String, dynamic> json) {
+    return Keyword(
         id: json['id'] as int?,
-        name: json['name'] as String,
+        name: json['name'] as String?,
+        label: json['label'] as String?,
         description: json['description'] as String?
     );
   }
@@ -41,9 +47,10 @@ class Unit {
   Map<String, dynamic> toJson() => {
     'id': this.id,
     'name': this.name,
+    'label': this.label,
     'description': this.description ?? ''
   };
 
   @override
-  String toString() => name;
+  String toString() => name ?? label ?? '';
 }
