@@ -3,6 +3,7 @@ import 'dart:convert';
 import 'package:http/http.dart';
 import 'package:tare/exceptions/api_exception.dart';
 import 'package:tare/models/user.dart';
+import 'package:tare/models/userToken.dart';
 import 'package:tare/models/user_setting.dart';
 import 'package:tare/services/api/api_service.dart';
 
@@ -16,14 +17,35 @@ class ApiUser extends ApiService {
       List<dynamic> json = jsonDecode(utf8.decode(res.bodyBytes));
       List<User> userList = [];
 
-      json.forEach((element) {
+      for (var element in json) {
         userList.add(User.fromJson(element));
-      });
+      }
 
       return userList;
     } else {
       throw ApiException(
           message: 'User api error - could not fetch user data',
+          statusCode: res.statusCode
+      );
+    }
+  }
+
+  Future<UserToken> createAuthToken(String username, String password) async {
+    var url = '/api-token-auth/';
+
+    Map<String, dynamic> requestData = {
+      'username': username,
+      'password': password
+    };
+
+    Response res = await httpPost(url, requestData, withoutToken: true);
+
+    if ([200, 201].contains(res.statusCode)) {
+      Map<String, dynamic> json = jsonDecode(utf8.decode(res.bodyBytes));
+      return UserToken.fromJson(json);
+    } else {
+      throw ApiException(
+          message: 'User api error - could not fetch user settings',
           statusCode: res.statusCode
       );
     }
